Add files via upload
Two major components of GANs are Generator and Discriminator. The role of the generator is like a thief to generate the fake samples based on the original sample and make the discriminator fool to understand Fake as real. On the other hand, a Discriminator is like a Police whose role is to identify the abnormalities in the samples created by Generator and classify them as Fake or real. This competition between both the component goes on until the level of perfection is achieved where Generator wins making a Discriminator fool on fake data. Applications of Generative Adversarial Networks (GANs) Reading about GANs is too exciting and when you will read their application then I hope that excitement will reach the sky and then study the working of GANs creates a different impact on learning.

Generate new data from available data – It means generating new samples from an available sample that is not similar to a real one. Generate realistic pictures of people that have never existed. Gans is not limited to Images, It can generate text, articles, songs, poems, etc. Generate Music by using some clone Voice – If you provide some voice then GANs can generate a similar clone feature of it. In this research paper, researchers from NIT in Tokyo proposed a system that is able to generate melodies from lyrics with help of learned relationships between notes and subjects. Text to Image Generation (Object GAN and Object Driven GAN) Creation of anime characters in Game Development and animation production. Image to Image Translation – We can translate one Image to another without changing the background of the source image. For example, Gans can replace a dog with a cat. Low resolution to High resolution – If you pass a low-resolution Image or video, GAN can produce a high-resolution Image version of the same. Prediction of Next Frame in the video – By training a neural network on small frames of video, GANs are capable to generate or predict a small next frame of video. For example, you can have a look at below GIF Interactive Image Generation – It means that GANs are capable to generate images and video footage in an art form if they are trained on the right real dataset. Speech – Researchers from the College of London recently published a system called GAN-TTS that learns to generate raw audio through training on 567 corpora of speech data.